+++ pkg/project/project.go
@@ -5,7 +5,7 @@ var (
 	gitSHA      = "n/a"
 	name        = "search-mcp"
 	source      = "https://github.com/giantswarm/search-mcp"
-	version     = "0.1.1-dev"
+	version     = "0.2.0"
 )
 
 func Description() string {